home rows refactor
