+++ tv/domain/src/main/java/com/rootstrap/domain/HomeRow.kt
@@ -1,0 +1,3 @@
+package com.rootstrap.domain
+
+data class HomeRow(val name: String, val rowItems: List<Movie>)
