+++ tv/src/main/java/com/rootstrap/tv/pages/home/HomeScreenViewModel.kt
@@ -6,19 +6,14 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.rootstrap.domain.Movie
 import com.rootstrap.tv.data.MoviesRepository
+import com.rootstrap.tv.models.HomeRowModel
+import com.rootstrap.tv.utils.HomeRowMapper
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 
 class HomeScreenViewModel(private val videoRepository: MoviesRepository) : ViewModel() {
-
-    companion object {
-        const val TOP_TEN = "Top 10 Movies"
-        const val NOW_PLAYING = "Now playing"
-        const val SCIENCE_FICTION = "Science fiction"
-    }
-
     private val uiState: MovieUiState =
         MovieUiState(SnapshotStateList(), SnapshotStateList())
     private var _uiStateFlow: MutableStateFlow<MovieUiState> =
@@ -29,25 +24,10 @@ class HomeScreenViewModel(private val videoRepository: MoviesRepository) : ViewM
 
     fun onHomeScreenLoaded() {
         viewModelScope.launch {
-            val topTenRow = HomeRow(
-                name = TOP_TEN,
-                rowItems = videoRepository.getTop10Movies().toMutableStateList()
-            )
-            val nowPlayingMovies = HomeRow(
-                name = NOW_PLAYING,
-                rowItems = videoRepository.getNowPlayingMovies().toMutableStateList()
-            )
-            val scienceFictionMovies = HomeRow(
-                name = SCIENCE_FICTION,
-                rowItems = videoRepository.getScienceFictionMovies().toMutableStateList()
-            )
+            val rows = HomeRowMapper.getHomeRows(videoRepository.getHomeRows())
             _uiStateFlow.update {
                 uiState.copy(
-                    homeRows = listOf(
-                        topTenRow,
-                        nowPlayingMovies,
-                        scienceFictionMovies
-                    ).toMutableStateList(),
+                    homeRows = rows,
                     featuredMovies = videoRepository.getFeaturedMovies().toMutableStateList()
                 )
             }
@@ -56,6 +36,6 @@ class HomeScreenViewModel(private val videoRepository: MoviesRepository) : ViewM
 }
 
 data class MovieUiState(
-    val homeRows: SnapshotStateList<HomeRow>,
+    val homeRows: SnapshotStateList<HomeRowModel>,
     val featuredMovies: SnapshotStateList<Movie>
 )
